installsh: Support absolute path to prepare_file()
Before it is assumed that path to file to prepare is relative to ROOT
directory. However in some cases we pass absolute path based prefixed
with ROOT that prefixed inside function again causing incorrect path
to be present.

Check if path to file is absolute and subpath under ROOT or relative
and prefixed internally with ROOT like in previous implementation.

While there check if non-existing destination is a broken symlink and
remove it in that case.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>