installsh: Fix destination prefix and directory creation
When source prefix is a directory source path extended with wildcards to
match all entries in that path. When source (and thus destination) is a
top level directory like /sbin or /etc and containing only non-directory
entries $MKDIR will not be executed as parent directory is a install
root directory that always exist.

All source entries installed as file like sbin or etc, making backups if
they are enabled.

Fix by suffixing destination path with slash when source is a directory.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>